pyproject.toml
[tool.black]
target-version = ["py38"]

[tool.poetry]
name = "simret"
version = "0.1.0"
description = "Framework for SimCLR pretrained Object Detection"
authors = [
    "Lars Vagnes <larsvagnes@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.7"
loguru = "0.5.*"
toml = "^0.10"
scikit-learn = "0.*"
scipy = "1.5.*"
pyyaml = "5.3.*"
tensorboard =  {version = "2.1", optional = true}
matplotlib = {version = "3.3.*", optional = true}
opencv-python = {version = "4.4.*", optional = true}
torch = "^1.8.1"
torchvision = "^0.9.1"
fire = "^0.4.0"


[tool.poetry.extras]
train = [
"tensorboard",
"matplotlib",
"opencv-python",
]

[tool.poetry.dev-dependencies]
black = "20.8b1"
flake8 = "^3.8"
ipython = "^7.18"
jedi = "0.17"  # jedi 0.18 crashes IPython
mypy = "^0.782"
pre-commit = "^2.7.1"
pytest = "^6.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
